Add extra install location for anaconda on Mac M1

+++ setup_Mac/STEP1-install_anaconda.command
@@ -25,6 +25,13 @@ if test -f $HOME/anaconda/bin/activate; then
     exit 0
 fi
 
+if test -f $HOME/opt/anaconda3/bin/activate; then
+    source $HOME/optanaconda/bin/activate
+    conda init/
+    conda init zsh
+    exit 0
+fi
+
 curl -o anaconda.sh https://repo.anaconda.com/archive/Anaconda3-2021.05-MacOSX-x86_64.sh
 bash anaconda.sh -b -y -p $HOME/anaconda
 source $HOME/anaconda/bin/activate
